feat: function to handle status code and error message specific to user creation
